// src/http/models/getContractById200.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * onec
 * OpenAPI spec version: 1.0.0
 */

export type GetContractById200 = {
  id: string;
  /** @nullable */
  city: string | null;
  /** @nullable */
  client: string | null;
  /** @nullable */
  state: string | null;
  /** @nullable */
  cnpj: string | null;
  /** @nullable */
  sindic: string | null;
  /** @nullable */
  year: string | null;
  /** @nullable */
  matter: string | null;
  /** @nullable */
  forecast: string | null;
  /** @nullable */
  contractTotal: string | null;
  /** @nullable */
  percentage: number | null;
  /** @nullable */
  signedContract: string | null;
  /** @nullable */
  status: string | null;
  /** @nullable */
  averageGuide: number | null;
  /** @nullable */
  partner: string | null;
  /** @nullable */
  partnerCommission: number | null;
  /** @nullable */
  counter: string | null;
  /** @nullable */
  email: string | null;
};